Password encode
codificar la pass cuando se crea el user y validar que no se vuelva a guardar cuando se edita el user
