+++ src/main/java/com/seminario/sleepingMotorhome/Utils/Passgenerator.java
@@ -5,13 +5,13 @@ import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 public class Passgenerator {
 
     //public static void main(String ...args) {
-    //    BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder(4);
-
-        //El String que mandamos al metodo encode es el password que queremos encriptar.
+    //   BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder(4);
+//
+    //    //El String que mandamos al metodo encode es el password que queremos encriptar.
     //    System.out.println(bCryptPasswordEncoder.encode("1234"));
-
-        //    $2a$04$Sx7mnuh2WwjQH67pi4j1PepFBLVfe6AchnaDJqJoNbmpPPRBTUvRG
-
+//
+    //        //$2a$04$Sx7mnuh2WwjQH67pi4j1PepFBLVfe6AchnaDJqJoNbmpPPRBTUvRG
+//
     //}
